golioth: lightdb: split out as new module
Move lightdb related functions out of golioth.c module, in order to get
better code separation and readability. This will also make more space for
low-level packet handling in golioth.c module.
